apps/etl/pyproject.toml
[project]
name = "etl"
version = "0.1.0"
description = "ETL pipeline for TripGuard Thailand"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "httpx>=0.27.0",
    "psycopg2-binary>=2.9.9",
    "neo4j>=5.22.0",
    "python-dotenv>=1.0.1",
    "pydantic>=2.7.0",
    "schedule>=1.2.2",
]

[project.scripts]
etl = "run_all:run_pipeline"

[tool.uv]
package = false
